Move dispatcher into calc graph package.
