+++ go/felix/calc/calc_graph_test.go
@@ -24,7 +24,6 @@ import (
 	"github.com/projectcalico/felix/go/datastructures/set"
 	"github.com/projectcalico/felix/go/felix/config"
 	"github.com/projectcalico/felix/go/felix/proto"
-	"github.com/projectcalico/felix/go/felix/store"
 	"github.com/projectcalico/libcalico-go/lib/backend/api"
 	. "github.com/projectcalico/libcalico-go/lib/backend/model"
 	"github.com/projectcalico/libcalico-go/lib/net"
@@ -737,7 +736,7 @@ const (
 
 func doStateSequenceTest(expandedTest StateList, flushStrategy flushStrategy) {
 	var validationFilter *ValidationFilter
-	var calcGraph *store.Dispatcher
+	var calcGraph *Dispatcher
 	var tracker *stateTracker
 	var eventBuf *EventBuffer
 	var lastState State
